Move all css to root of the package
