--- packages/opensrp-plans/src/containers/PlanAssignment/index.css
@@ -1,5 +1,0 @@
-.plan-detail-view {
-  background-color: #ffffff;
-  min-height: 156px;
-  border-bottom: 1px solid #e9e9e9;
-}
