+++ packages/opensrp-plans/src/index.css
@@ -13,3 +13,82 @@
   padding: 24px;
   background: #fff;
 }
+
+.plan-detail-view {
+  background-color: #ffffff;
+  min-height: 156px;
+  border-bottom: 1px solid #e9e9e9;
+}
+
+.plan-info-main {
+  display: inline-flex;
+  align-items: center;
+  padding-left: 29px;
+  justify-content: space-between;
+}
+.plan-detail-view .site-page-header-responsive {
+  line-height: 22px;
+  padding-top: 16px;
+  padding-left: 36px;
+}
+.plan-title {
+  margin-left: 19px;
+}
+.plan-title h4 {
+  color: rgba(0, 0, 0, 0.85);
+  font-size: 20px;
+  line-height: 28px;
+  text-align: left;
+}
+.plan-title a,
+.plan-title span {
+  font-size: 14px;
+  line-height: 22px;
+  text-align: left;
+}
+.plan-start {
+  margin-left: 474px;
+}
+.plan-start-title {
+  margin-left: 6.9rem;
+}
+.plan-end-title {
+  margin-left: 6.9rem;
+}
+.plan-start-title,
+.plan-end-title {
+  color: rgba(0, 0, 0, 0.45);
+  font-size: 14px;
+  line-height: 36px;
+  text-align: right;
+  margin-left: 7.9rem;
+}
+.plan-info-main h2 {
+  color: rgba(0, 0, 0, 0.85);
+  font-size: 30px;
+  line-height: 38px;
+  text-align: right;
+}
+.plan-end {
+  margin-left: 14px;
+}
+.plan-info-main .timing-period-divider {
+  height: 40px;
+  margin-left: 31px;
+  margin-bottom: 20px;
+}
+@media only screen and (max-width: 1600px) {
+  .plan-start {
+    margin-left: 74px;
+  }
+}
+@media only screen and (max-width: 1200px) {
+  .plan-start {
+    margin-left: -11px;
+  }
+}
+@media only screen and (max-width: 1107px) {
+  .plan-info-main {
+    flex-direction: column;
+  }
+}
